--- srad-app/src/generic.rs
@@ -1,1 +1,0 @@
-mod application;
